Migrating polis methods for reports to wikipollservice and fixing spelling
