+++ api/src/tools/polis.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashMap, sync::Arc};
+use std::sync::Arc;
 
 use aide::axum::{
     routing::{get_with, post_with},
@@ -19,7 +19,7 @@ use uuid::Uuid;
 use crate::{
     error::ComhairleError,
     models,
-    wiki_poll_service::{WikiPollLogin, WikiPollService},
+    wiki_poll_service::{polis_service::WikiPollReport, WikiPollLogin, WikiPollService},
     ComhairleState,
 };
 
@@ -35,7 +35,7 @@ pub struct PolisToolConfig {
 }
 
 impl ToolConfigSanitize for PolisToolConfig {
-    fn sanatize(&self) -> Self {
+    fn sanitize(&self) -> Self {
         Self {
             admin_user: "".into(),
             admin_password: "".into(),
@@ -52,65 +52,6 @@ pub struct PolisToolSetup {
 
 #[derive(Clone, Serialize, Deserialize, Debug, JsonSchema)]
 pub struct PolisReport;
-
-// Report data structures
-#[derive(Serialize, Deserialize, Debug, JsonSchema)]
-pub struct CommentReportData {
-    pub tid: u32,
-    pub text: String,
-    pub overall_votes: VoteCounts,
-    pub group_votes: Vec<GroupVoteCounts>,
-    pub group_informed_consensus: Option<f64>,
-    pub divisiveness: Option<f64>,
-}
-
-#[derive(Serialize, Deserialize, Debug, JsonSchema, Clone)]
-pub struct VoteCounts {
-    pub agrees: u32,
-    pub disagrees: u32,
-    pub passes: u32,
-}
-
-#[derive(Serialize, Deserialize, Debug, JsonSchema)]
-pub struct GroupVoteCounts {
-    pub group_id: u32,
-    pub agrees: u32,
-    pub disagrees: u32,
-    pub passes: u32,
-}
-
-#[derive(Serialize, Deserialize, Debug, JsonSchema)]
-pub struct GroupReportData {
-    pub group_id: u32,
-    pub representative_comments: Vec<RepresentativeComment>,
-    pub members: Vec<u32>, // pids
-}
-
-#[derive(Serialize, Deserialize, Debug, JsonSchema)]
-pub struct RepresentativeComment {
-    pub tid: u32,
-    pub text: String,
-}
-
-#[derive(Serialize, Deserialize, Debug, JsonSchema)]
-pub struct ParticipantReportData {
-    pub pid: u32,
-    pub group_id: Option<u32>,
-    pub pca_position: Option<PcaPosition>,
-}
-
-#[derive(Serialize, Deserialize, Debug, JsonSchema)]
-pub struct PcaPosition {
-    pub x: f64,
-    pub y: f64,
-}
-
-#[derive(Serialize, Deserialize, Debug, JsonSchema)]
-pub struct PolisReportResponse {
-    pub comments: Vec<CommentReportData>,
-    pub groups: Vec<GroupReportData>,
-    pub participants: Vec<ParticipantReportData>,
-}
 
 /// Zero-sized marker type for Polis tool implementation
 pub struct PolisTool;
@@ -138,7 +79,7 @@ impl ToolImpl for PolisTool {
     }
 
     fn sanitize(config: Self::Config) -> Self::Config {
-        config.sanatize()
+        config.sanitize()
     }
 
     fn routes(state: &Arc<ComhairleState>) -> ApiRouter {
@@ -222,18 +163,6 @@ pub struct PolisClient {
 }
 
 #[derive(Deserialize, Serialize, Debug)]
-pub struct PolisComment {
-    pub tid: u32,
-    pub txt: String,
-    pub is_seed: bool,
-    pub is_meta: bool,
-    pub lang: Option<String>,
-    pub pid: u32,
-    pub quote_src_url: Option<String>,
-    pub created: String,
-}
-
-#[derive(Deserialize, Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct PolisCommentCreateResponse {
     tid: u8,
@@ -253,459 +182,6 @@ pub struct SetTopicRequest {
     pub conversation_id: String,
 }
 
-impl PolisClient {
-    pub fn new(base_url: &str) -> Self {
-        let client = Client::builder().cookie_store(true).build().unwrap();
-        Self {
-            client,
-            base_url: base_url.to_string(),
-        }
-    }
-
-    pub async fn create_random_admin_user(&self) -> Result<(String, String), PolisError> {
-        info!("Creating a random admin user");
-        let username: String = rand::thread_rng()
-            .sample_iter(&Alphanumeric)
-            .take(10)
-            .map(char::from)
-            .collect();
-
-        let email = format!("{username}@comhairle.com");
-
-        let password: String = rand::thread_rng()
-            .sample_iter(&Alphanumeric)
-            .take(10)
-            .map(char::from)
-            .collect();
-
-        let new_user = NewAdminUser {
-            hname: username.clone(),
-            password: password.clone(),
-            password2: password,
-            email,
-            gatekeeperTosPrivacy: true,
-        };
-
-        let _res = self
-            .client
-            .post(format!("https://{}/api/v3/auth/new", self.base_url))
-            .json(&new_user)
-            .send()
-            .await
-            .map_err(|e| {
-                warn!("{e}");
-                PolisError::FailedToCreateNewAdminUser
-            })?
-            .text()
-            // .json::<NewUserResp>()
-            .await
-            .map_err(|e| {
-                warn!("{e}");
-                PolisError::FailedToCreateNewAdminUser
-            })?;
-
-        Ok((new_user.email, new_user.password))
-    }
-
-    async fn login(&self, login: &PolisLogin) -> Result<String, PolisError> {
-        info!("Logging in to polis");
-        let url = format!("https://{}/api/v3/auth/login", self.base_url);
-        println!("format {url}");
-        let resp = self
-            .client
-            .post(url)
-            .json(&login)
-            .send()
-            .await
-            .map_err(|e| {
-                println!("First bit {e}");
-                PolisError::FailedToLogin
-            })?;
-
-        let cookie = resp
-            .headers()
-            .get(SET_COOKIE)
-            .ok_or(PolisError::FailedToLogin)?
-            .to_str()
-            .map_err(|_| PolisError::FailedToLogin)?
-            .to_owned();
-
-        let login_resp = resp
-            .json::<LoginResp>()
-            // .text()
-            .await
-            .map_err(|e| {
-                println!("{e}");
-                PolisError::FailedToLogin
-            })?;
-
-        info!("Logged user into polis {login_resp:#?}");
-
-        Ok(cookie)
-    }
-
-    pub async fn create_poll(&self) -> Result<String, PolisError> {
-        info!("Attempting to create a new poll");
-        let new_poll = self
-            .client
-            .post(format!("https://{}/api/v3/conversations", self.base_url))
-            .send()
-            .await
-            .map_err(|e| {
-                warn!("Failed to create new poll: {e:#?}");
-                PolisError::FailedToCreateNewPoll
-            })?
-            // .text()
-            .json::<NewPollResp>()
-            .await
-            .map_err(|e| {
-                warn!("Failed to create new poll: {e:#?}");
-                PolisError::FailedToCreateNewPoll
-            })?;
-        Ok(new_poll.conversation_id.to_owned())
-    }
-
-    pub async fn set_topic(&self, topic: SetTopicRequest) -> Result<(), PolisError> {
-        let body = self
-            .client
-            .put(format!("{}/api/v3/conversations", self.base_url))
-            .json(&topic)
-            .send()
-            .await
-            .unwrap()
-            .text()
-            .await
-            .unwrap();
-
-        println!("{body}");
-        Ok(())
-    }
-
-    #[instrument(err(Debug), skip(self))]
-    pub async fn post_seed_comment(
-        &self,
-        comment: &str,
-        poll_id: &str,
-    ) -> Result<String, PolisError> {
-        let post_json =
-            json!({"txt":comment,"pid":"mypid","conversation_id":poll_id,"is_seed":true});
-
-        let resp = self
-            .client
-            .post(format!("{}/api/v3/comments", self.base_url))
-            .json(&post_json)
-            .send()
-            .await
-            .map_err(|e| PolisError::FailedToPostSeedComment(e.to_string()))?
-            .json::<PolisCommentCreateResponse>()
-            .await
-            .map_err(|e| PolisError::FailedToPostSeedComment(e.to_string()))?;
-
-        Ok(resp.tid.to_string())
-    }
-
-    pub async fn get_comments(&self, poll_id: &str) -> Result<Vec<PolisComment>, PolisError> {
-        let url = format!(
-            "{}/api/v3/comments?conversation_id={poll_id}",
-            self.base_url
-        );
-        let comments: Vec<PolisComment> =
-            self.client
-                .get(url)
-                .send()
-                .await
-                .unwrap()
-                .json()
-                .await
-                .map_err(|e| PolisError::FailedToGetComments(e.to_string()))?;
-
-        Ok(comments)
-    }
-
-    pub async fn get_math_pca(&self, poll_id: &str) -> Result<serde_json::Value, PolisError> {
-        let url = format!(
-            "https://{}/api/v3/math/pca2?conversation_id={}&lastVoteTimestamp=0",
-            self.base_url, poll_id
-        );
-
-        info!("Getting PCA data from {url}");
-
-        let response = self.client.get(&url).send().await.map_err(|e| {
-            warn!("Failed to get PCA data: {e}");
-            PolisError::FailedToGetComments(format!("Failed to get PCA data: {e}"))
-        })?;
-
-        info!("Response {response:#?}");
-        let data = response.json::<serde_json::Value>().await.map_err(|e| {
-            warn!("Failed to parse PCA data: {e:#?}");
-            PolisError::FailedToGetComments(format!("Failed to parse PCA data: {e}"))
-        })?;
-
-        Ok(data)
-    }
-
-    pub async fn get_conversation(&self, poll_id: &str) -> Result<serde_json::Value, PolisError> {
-        let url = format!(
-            "https://{}/api/v3/conversations?conversation_id={}",
-            self.base_url, poll_id
-        );
-
-        let response = self.client.get(&url).send().await.map_err(|e| {
-            warn!("Failed to get conversation: {e}");
-            PolisError::FailedToGetComments(format!("Failed to get conversation: {e}"))
-        })?;
-
-        let data = response.json::<serde_json::Value>().await.map_err(|e| {
-            warn!("Failed to parse conversation: {e}");
-            PolisError::FailedToGetComments(format!("Failed to parse conversation: {e}"))
-        })?;
-
-        Ok(data)
-    }
-
-    pub async fn get_comments_with_voting(
-        &self,
-        poll_id: &str,
-    ) -> Result<serde_json::Value, PolisError> {
-        let url = format!(
-            "https://{}/api/v3/comments?conversation_id={}&moderation=true&include_voting_patterns=true",
-            self.base_url, poll_id
-        );
-
-        let response = self.client.get(&url).send().await.map_err(|e| {
-            warn!("Failed to get comments: {e}");
-            PolisError::FailedToGetComments(format!("Failed to get comments: {e}"))
-        })?;
-
-        let data = response.json::<serde_json::Value>().await.map_err(|e| {
-            warn!("Failed to parse comments: {e}");
-            PolisError::FailedToGetComments(format!("Failed to parse comments: {e}"))
-        })?;
-
-        Ok(data)
-    }
-
-    pub async fn get_report_data(&self, poll_id: &str) -> Result<PolisReportResponse, PolisError> {
-        info!("Getting full report data for poll_id: {poll_id}");
-
-        // Fetch all the data that powers the report page
-        let math_pca = self.get_math_pca(poll_id).await?;
-        let comments_data = self.get_comments_with_voting(poll_id).await?;
-
-        info!("math pca {math_pca:#?}");
-        info!("comments {comments_data:#?}");
-
-        // Transform the raw data into structured report format
-        self.transform_report_data(math_pca, comments_data)
-    }
-
-    fn transform_report_data(
-        &self,
-        math_pca: serde_json::Value,
-        comments_data: serde_json::Value,
-    ) -> Result<PolisReportResponse, PolisError> {
-        // Extract comment texts from comments_data
-        let comments_array = comments_data
-            .as_array()
-            .ok_or_else(|| PolisError::FailedToGetComments("Invalid comments format".into()))?;
-
-        // Create a map of tid -> comment text
-        let mut comment_texts = std::collections::HashMap::new();
-        let mut comment_votes: HashMap<u32, VoteCounts> = std::collections::HashMap::new();
-
-        for comment in comments_array {
-            if let (Some(tid), Some(txt), agrees, disagrees, passes) = (
-                comment.get("tid").and_then(|t| t.as_u64()),
-                comment.get("txt").and_then(|t| t.as_str()),
-                comment
-                    .get("agree_count")
-                    .and_then(|t| t.as_u64())
-                    .unwrap_or(0) as u32,
-                comment
-                    .get("disagree_count")
-                    .and_then(|t| t.as_u64())
-                    .unwrap_or(0) as u32,
-                comment
-                    .get("pass_count")
-                    .and_then(|t| t.as_u64())
-                    .unwrap_or(0) as u32,
-            ) {
-                comment_texts.insert(tid as u32, txt.to_string());
-
-                comment_votes.insert(
-                    tid as u32,
-                    VoteCounts {
-                        agrees,
-                        disagrees,
-                        passes,
-                    },
-                );
-            }
-        }
-
-        // Extract data from math_pca
-        let tids = math_pca["tids"]
-            .as_array()
-            .ok_or_else(|| PolisError::FailedToGetComments("No tids in math data".into()))?;
-
-        let group_votes = &math_pca["group-votes"];
-        let group_aware_consensus = &math_pca["group-aware-consensus"];
-        let empty_vec = vec![];
-        let comment_extremity = math_pca["pca"]["comment-extremity"]
-            .as_array()
-            .unwrap_or(&empty_vec);
-
-        // Build comments with vote counts
-        let mut comments_report = Vec::new();
-        for (idx, tid_val) in tids.iter().enumerate() {
-            let tid = tid_val.as_u64().unwrap_or(0) as u32;
-            let text = comment_texts.get(&tid).cloned().unwrap_or_default();
-
-            // Calculate overall votes from group votes
-            let mut group_votes_list = Vec::new();
-
-            if let Some(obj) = group_votes.as_object() {
-                for (group_id_str, group_vote_data) in obj.iter() {
-                    if let Ok(group_id) = group_id_str.parse::<u32>() {
-                        if let Some(votes_for_tid) = group_vote_data.get(tid.to_string()) {
-                            let agrees = votes_for_tid["A"].as_u64().unwrap_or(0) as u32;
-                            let disagrees = votes_for_tid["D"].as_u64().unwrap_or(0) as u32;
-                            let passes = votes_for_tid["S"].as_u64().unwrap_or(0) as u32;
-                            group_votes_list.push(GroupVoteCounts {
-                                group_id,
-                                agrees,
-                                disagrees,
-                                passes,
-                            });
-                        }
-                    }
-                }
-            }
-
-            let consensus = group_aware_consensus
-                .get(tid.to_string())
-                .and_then(|v| v.as_f64());
-
-            let divisiveness = comment_extremity.get(idx).and_then(|v| v.as_f64());
-
-            comments_report.push(CommentReportData {
-                tid,
-                text,
-                overall_votes: (*comment_votes.get(&tid).unwrap()).clone(),
-                group_votes: group_votes_list,
-                group_informed_consensus: consensus,
-                divisiveness,
-            });
-        }
-
-        // Extract group clusters and build groups
-        let empty_clusters = vec![];
-        let group_clusters = math_pca["group-clusters"]
-            .as_array()
-            .unwrap_or(&empty_clusters);
-
-        let repness = &math_pca["repness"];
-
-        let mut groups_report = Vec::new();
-        for (idx, cluster) in group_clusters.iter().enumerate() {
-            let group_id = idx as u32;
-            let members: Vec<u32> = cluster["members"]
-                .as_array()
-                .map(|arr| {
-                    arr.iter()
-                        .filter_map(|v| v.as_u64().map(|n| n as u32))
-                        .collect()
-                })
-                .unwrap_or_default();
-
-            // Get representative comments for this group
-            let mut representative_comments = Vec::new();
-            if let Some(rep_array) = repness.get(group_id.to_string()).and_then(|v| v.as_array()) {
-                for rep in rep_array {
-                    if let Some(tid) = rep.get("tid").and_then(|t| t.as_u64()) {
-                        let tid = tid as u32;
-                        let text = comment_texts.get(&tid).cloned().unwrap_or_default();
-                        representative_comments.push(RepresentativeComment { tid, text });
-                    }
-                }
-            }
-
-            groups_report.push(GroupReportData {
-                group_id,
-                representative_comments,
-                members,
-            });
-        }
-
-        // Build participants list with group membership
-        let bid_to_pid = &math_pca["bidToPid"];
-        let pca_x = math_pca["pca"]["x"].as_array();
-        let pca_y = math_pca["pca"]["y"].as_array();
-
-        let mut participants_report = Vec::new();
-        let mut pid_to_group: std::collections::HashMap<u32, u32> =
-            std::collections::HashMap::new();
-
-        // Build pid to group mapping from group clusters
-        for (group_id, cluster) in group_clusters.iter().enumerate() {
-            if let Some(members) = cluster["members"].as_array() {
-                for member in members {
-                    if let Some(pid) = member.as_u64() {
-                        pid_to_group.insert(pid as u32, group_id as u32);
-                    }
-                }
-            }
-        }
-
-        // If we have bidToPid mapping, use it to get all participants
-        if let Some(bid_array) = bid_to_pid.as_array() {
-            for (bid, pids_value) in bid_array.iter().enumerate() {
-                if let Some(pids) = pids_value.as_array() {
-                    for pid_val in pids {
-                        if let Some(pid) = pid_val.as_u64() {
-                            let pid = pid as u32;
-                            let group_id = pid_to_group.get(&pid).copied();
-
-                            // Get PCA position if available
-                            let pca_position = if let (Some(x_arr), Some(y_arr)) = (pca_x, pca_y) {
-                                x_arr.get(bid).and_then(|x| x.as_f64()).and_then(|x| {
-                                    y_arr
-                                        .get(bid)
-                                        .and_then(|y| y.as_f64())
-                                        .map(|y| PcaPosition { x, y })
-                                })
-                            } else {
-                                None
-                            };
-
-                            participants_report.push(ParticipantReportData {
-                                pid,
-                                group_id,
-                                pca_position,
-                            });
-                        }
-                    }
-                }
-            }
-        } else {
-            // Fallback: just list participants from groups
-            for (pid, group_id) in pid_to_group.iter() {
-                participants_report.push(ParticipantReportData {
-                    pid: *pid,
-                    group_id: Some(*group_id),
-                    pca_position: None,
-                });
-            }
-        }
-
-        Ok(PolisReportResponse {
-            comments: comments_report,
-            groups: groups_report,
-            participants: participants_report,
-        })
-    }
-}
-
 #[derive(Serialize, Deserialize, JsonSchema)]
 struct AdminLoginQuery {
     pub workflow_step_id: Uuid,
@@ -720,7 +196,7 @@ struct ReportDataQuery {
 async fn get_report_data(
     State(state): State<Arc<ComhairleState>>,
     Query(ReportDataQuery { workflow_step_id }): Query<ReportDataQuery>,
-) -> Result<(StatusCode, Json<PolisReportResponse>), ComhairleError> {
+) -> Result<(StatusCode, Json<WikiPollReport>), ComhairleError> {
     let workflow_step = models::workflow_step::get_by_id(&state.db, &workflow_step_id).await?;
 
     println!("{workflow_step:#?}");
@@ -731,18 +207,20 @@ async fn get_report_data(
         _ => return Err(ComhairleError::WorkflowStepHasWrongType("Polis".into())),
     };
 
-    let client = PolisClient::new(&config.server_url);
-
     // Login as admin to access the data export
-    client
-        .login(&PolisLogin {
+    state
+        .wiki_poll_service
+        .login(&WikiPollLogin {
             email: config.admin_user,
             password: config.admin_password,
         })
         .await?;
 
     // Get the report data
-    let data = client.get_report_data(&config.poll_id).await?;
+    let data = state
+        .wiki_poll_service
+        .get_report_data(&config.poll_id)
+        .await?;
 
     Ok((StatusCode::OK, Json(data)))
 }
